requirements.txt
openai~=1.29.0
python-dotenv==1.0.0
anthropic~=0.25.9
